requirements.txt
# Core
python-hcl2>=4.3.0
requests>=2.31.0
pydantic>=2.5.0

# API Server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Database
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.25
alembic>=1.13.1

# CLI
rich>=13.7.0

# Templating
jinja2>=3.1.3

# AWS
boto3>=1.34.0

# PDF Reports
reportlab>=4.0.0

# Utilities
python-dotenv>=1.0.0

# Azure
azure-mgmt-resource>=23.0.0
azure-mgmt-compute>=30.0.0
azure-mgmt-network>=25.0.0
azure-mgmt-sql>=4.0.0b12
azure-identity>=1.15.0

# GCP
google-cloud-resource-manager>=1.12.0
google-cloud-compute>=1.15.0
google-cloud-storage>=2.14.0
google-auth>=2.27.0
